fix: blank connections & search bar
