+++ src/ui/pages/Connections/Connections.scss
@@ -40,7 +40,7 @@
   }
 
   .page-header {
-    height: calc(var(--ion-safe-area-top, 0) + 3.75rem);
+    height: calc(var(--ion-safe-area-top, 0) + 4.25rem);
     transition: all 0.25s;
 
     ion-toolbar {
@@ -49,21 +49,38 @@
     }
   }
 
+  .connections-header-area {
+    display: flex;
+    flex-direction: column;
+    background: var(--ion-color-neutral-200);
+    z-index: 2;
+    position: relative;
+  }
+
   &.hide-header {
     .page-header {
       height: var(--ion-safe-area-top, 0);
       --min-height: 0;
+      overflow: hidden;
+      opacity: 0;
+      pointer-events: none;
     }
 
-    .connections-list {
-      height: calc(100vh - (3.5rem + env(safe-area-inset-bottom)));
+    .connection-search-input {
+      opacity: 1 !important;
+      pointer-events: auto !important;
     }
   }
 
   .connections-body {
+    height: 100%;
     display: flex;
     flex-direction: row;
     margin-bottom: 0 !important;
+
+    .connections-list {
+      height: 100%;
+    }
   }
 
   .alphabet-selector {
