+++ src/ui/pages/Connections/components/ConnectionsBody/ConnectionsBody.tsx
@@ -8,58 +8,22 @@ import {
   IonItemGroup,
   IonLabel,
   IonRow,
-  IonSearchbar,
 } from "@ionic/react";
 import { useEffect, useRef, useState } from "react";
-import { i18n } from "../../../../../i18n";
 import { combineClassNames } from "../../../../utils/style";
 import { AlphabetSelector } from "../AlphabetSelector";
 import { AlphabeticList } from "../AlphabeticList";
 
 import { SearchConnectionContent } from "../SearchConnectionContent";
 import "./ConnectionsBody.scss";
-import {
-  ConnectionsBodyProps,
-  SearchInputProps,
-} from "./ConnectionsBody.types";
-
-const SearchInput = ({ onFocus, onInputChange, value }: SearchInputProps) => {
-  const showCancel = value ? "always" : "focus";
-
-  const handleBlur = () => {
-    if (value) return;
-    onFocus?.(false);
-  };
-
-  const handleCancer = () => {
-    onFocus?.(false);
-  };
-
-  return (
-    <IonSearchbar
-      className="connection-search-input"
-      showCancelButton={showCancel}
-      onIonCancel={handleCancer}
-      debounce={100}
-      onIonFocus={() => onFocus?.(true)}
-      onIonBlur={handleBlur}
-      value={value}
-      onIonInput={(e) => {
-        onInputChange(e.target.value || "");
-      }}
-      placeholder={`${i18n.t("connections.page.search.placeholder")}`}
-    />
-  );
-};
+import { ConnectionsBodyProps } from "./ConnectionsBody.types";
 
 const ALPHABET_LIST_MAX_HEIGHT = 432;
 
 const ConnectionsBody = ({
   mappedConnections,
   handleShowConnectionDetails,
-  onSearchFocus,
   search,
-  setSearch,
 }: ConnectionsBodyProps) => {
   const [keyboardIsOpen, setKeyboardIsOpen] = useState(false);
   const container = useRef<HTMLDivElement>(null);
@@ -119,54 +83,47 @@ const ConnectionsBody = ({
   });
 
   return (
-    <>
-      <SearchInput
-        onInputChange={setSearch}
-        value={search}
-        onFocus={onSearchFocus}
-      />
-      <div
-        className={classes}
-        ref={container}
-      >
-        <IonContent className="connections-list">
-          <IonGrid>
-            <IonRow>
-              <IonCol size="12">
-                {!search &&
-                  mappedConnections.map((alphabeticGroup, index) => {
-                    return (
-                      <IonItemGroup
-                        className="connections-list-alphabetic-block"
-                        data-testid={`connection-group-${index}`}
-                        key={index}
-                      >
-                        <IonItemDivider id={alphabeticGroup.key}>
-                          <IonLabel>{alphabeticGroup.key}</IonLabel>
-                        </IonItemDivider>
-                        <AlphabeticList
-                          items={Array.from(alphabeticGroup.value)}
-                          handleShowConnectionDetails={
-                            handleShowConnectionDetails
-                          }
-                        />
-                      </IonItemGroup>
-                    );
-                  })}
-                {search && (
-                  <SearchConnectionContent
-                    keyword={search}
-                    mappedConnections={mappedConnections}
-                    onItemClick={handleShowConnectionDetails}
-                  />
-                )}
-              </IonCol>
-            </IonRow>
-          </IonGrid>
-        </IonContent>
-        {!search && <AlphabetSelector />}
-      </div>
-    </>
+    <div
+      className={classes}
+      ref={container}
+    >
+      <IonContent className="connections-list">
+        <IonGrid>
+          <IonRow>
+            <IonCol size="12">
+              {!search &&
+                mappedConnections.map((alphabeticGroup, index) => {
+                  return (
+                    <IonItemGroup
+                      className="connections-list-alphabetic-block"
+                      data-testid={`connection-group-${index}`}
+                      key={index}
+                    >
+                      <IonItemDivider id={alphabeticGroup.key}>
+                        <IonLabel>{alphabeticGroup.key}</IonLabel>
+                      </IonItemDivider>
+                      <AlphabeticList
+                        items={Array.from(alphabeticGroup.value)}
+                        handleShowConnectionDetails={
+                          handleShowConnectionDetails
+                        }
+                      />
+                    </IonItemGroup>
+                  );
+                })}
+              {search && (
+                <SearchConnectionContent
+                  keyword={search}
+                  mappedConnections={mappedConnections}
+                  onItemClick={handleShowConnectionDetails}
+                />
+              )}
+            </IonCol>
+          </IonRow>
+        </IonGrid>
+      </IonContent>
+      {!search && <AlphabetSelector />}
+    </div>
   );
 };
