+++ src/ui/pages/Connections/components/ConnectionsBody/ConnectionsBody.types.ts
@@ -1,12 +1,6 @@
 import { ConnectionShortDetails } from "../../../../../core/agent/agent.types";
 import { MappedConnections } from "../../Connections.types";
 
-interface SearchInputProps {
-  onFocus?: (value: boolean) => void;
-  value: string;
-  onInputChange: (value: string) => void;
-}
-
 interface ConnectionsBodyProps {
   onSearchFocus?: (value: boolean) => void;
   mappedConnections: MappedConnections[];
@@ -30,7 +24,6 @@ interface SearchConnectionContentProps {
 
 export type {
   ConnectionsBodyProps,
-  SearchInputProps,
   SearchConnectionListProps,
   SearchConnectionContentProps,
 };
