+++ src/ui/pages/Connections/components/SearchInput/SearchInput.types.ts
@@ -1,0 +1,7 @@
+interface SearchInputProps {
+  onFocus?: (value: boolean) => void;
+  value: string;
+  onInputChange: (value: string) => void;
+}
+
+export type { SearchInputProps };
