+++ src/ui/pages/Connections/components/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchInput";
